requirements.txt
numpy>=1.20
streamlit>=1.8.1
protobuf~=3.19.0
tensorflow-cpu==2.9.0
